fixed beaglebone breaking changes from raspi network integration

+++ scripts/delete_all_wifi_connections.sh
@@ -1,8 +1,28 @@
-#!/bin/sh
- 
-echo "Using sudo to configure your networking, please enter your password:"
-sudo service connman stop
-sleep 5
-sudo rm -fr /var/lib/connman/*.config /var/lib/connman/wifi_*
-sudo service connman start
- 
+#! /bin/bash
+
+# Check virtual environment is activated
+if [ -z "${VIRTUAL_ENV}" ] ; then
+    echo "Please activate your virtual environment then re-run script"
+    exit 0
+fi
+
+# Check platform info is sourced
+if [[ -z "$PLATFORM" ]]; then
+	source $OPENAG_BRAIN_ROOT/scripts/get_platform_info.sh
+fi
+
+# Delete all wifi connections
+echo "Deleting all wifi connections..."
+
+# Delete wifi connections for beaglebone black wireless
+if [[ $PLATFORM == "beaglebone-black-wireless" ]]; then
+	sudo service connman stop
+	sleep 3
+	sudo rm -fr /var/lib/connman/*.config /var/lib/connman/wifi_*
+	sudo service connman start
+
+# Invalid platform
+else
+	echo "Unable to delete all wifi connections, wifi not supported for $PLATFORM"
+	exit 0
+fi
